Move to ts tests
