+++ packages/code-du-travail-frontend/src/outils/DureePreavisDemission/steps/__tests__/Informations.test.tsx
